Refactor digital document and witness deposition sign modals to use shared components for modal header and close button; streamline payment modals by consolidating payment selection logic into a reusable component; enhance address comparison and user option retrieval for order channels; improve code organization and readability.
