+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/home/src/pages/employee/WitnessDepositionSignModal.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { CloseSvg, InfoCard } from "@egovernments/digit-ui-components";
+import { InfoCard } from "@egovernments/digit-ui-components";
 import { Button, Loader } from "@egovernments/digit-ui-react-components";
 import { FileDownloadIcon, FileUploadIcon } from "@egovernments/digit-ui-module-dristi/src/icons/svgIndex";
 import AuthenticatedLink from "@egovernments/digit-ui-module-dristi/src/Utils/authenticatedLink";
@@ -11,6 +11,7 @@ import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
 import { witnessDepositionWorkflowAction } from "@egovernments/digit-ui-module-dristi/src/Utils/submissionWorkflow";
 import { useLocation } from "react-router-dom/cjs/react-router-dom";
 import { SIGNATURE_UPLOAD_CONFIG, buildUploadModalConfig, UploadModal, getUploadErrorToast } from "@egovernments/digit-ui-module-common";
+import { SignModalCloseBtn, SignModalHeading } from "./shared/signModalChrome";
 
 export const clearWitnessDepositionSessionData = () => {
   sessionStorage.removeItem("esignProcess");
@@ -140,22 +141,6 @@ export const WitnessDepositionSignModal = ({
     filingNumber,
     docObj?.artifactList.artifactNumber,
   ]);
-
-  const CloseBtn = useCallback((props) => {
-    return (
-      <div onClick={props?.onClick} style={{ height: "100%", display: "flex", alignItems: "center", paddingRight: "20px", cursor: "pointer" }}>
-        <CloseSvg />
-      </div>
-    );
-  }, []);
-
-  const Heading = useCallback((props) => {
-    return (
-      <div className="evidence-title">
-        <h1 className="heading-m">{props.label}</h1>
-      </div>
-    );
-  }, []);
 
   const uploadModalConfig = useMemo(() => buildUploadModalConfig(name, SIGNATURE_UPLOAD_CONFIG), [name]);
 
@@ -420,9 +405,9 @@ export const WitnessDepositionSignModal = ({
       )}
       {stepper === 0 && (
         <Modal
-          headerBarEnd={<CloseBtn onClick={handleCancel} />}
+          headerBarEnd={<SignModalCloseBtn onClick={handleCancel} />}
           headerBarMain={
-            <Heading
+            <SignModalHeading
               label={
                 witnessDepositionLoader
                   ? t(" ")
@@ -457,8 +442,8 @@ export const WitnessDepositionSignModal = ({
       {/* to select e-sign or upload */}
       {stepper === 1 && !openUploadSignatureModal && !isSigned && (
         <Modal
-          headerBarMain={<Heading label={t("ADD_SIGNATURE")} />}
-          headerBarEnd={<CloseBtn onClick={handleCancel} />}
+          headerBarMain={<SignModalHeading label={t("ADD_SIGNATURE")} />}
+          headerBarEnd={<SignModalCloseBtn onClick={handleCancel} />}
           actionCancelLabel={t("CS_COMMON_BACK")}
           actionCancelOnSubmit={handleCancel}
           actionSaveLabel={t("CS_COMMON_SUBMIT")}
@@ -529,8 +514,8 @@ export const WitnessDepositionSignModal = ({
       {/* after signing showing signed modal */}
       {stepper === 1 && !openUploadSignatureModal && isSigned && (
         <Modal
-          headerBarMain={<Heading label={t("ADD_SIGNATURE")} />}
-          headerBarEnd={<CloseBtn onClick={handleCancel} />}
+          headerBarMain={<SignModalHeading label={t("ADD_SIGNATURE")} />}
+          headerBarEnd={<SignModalCloseBtn onClick={handleCancel} />}
           actionCancelLabel={t("CS_COMMON_BACK")}
           actionCancelOnSubmit={handleCancel}
           actionSaveLabel={t("SUBMIT_BUTTON")}
