+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/orders/src/pages/employee/PaymentForSummonModal.js
@@ -1,10 +1,8 @@
 import React, { useMemo, useState, useEffect } from "react";
-import { RadioButtons, CardLabel, LabelFieldPair } from "@egovernments/digit-ui-react-components";
 import { useTranslation } from "react-i18next";
 import CustomToast from "@egovernments/digit-ui-module-dristi/src/components/CustomToast";
 import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
-import { InfoCard, Loader } from "@egovernments/digit-ui-components";
-import ApplicationInfoComponent from "../../components/ApplicationInfoComponent";
+import { Loader } from "@egovernments/digit-ui-components";
 import DocumentModal from "../../components/DocumentModal";
 import usePaymentProcess from "../../../../home/src/hooks/usePaymentProcess";
 import { DRISTIService } from "@egovernments/digit-ui-module-dristi/src/services";
@@ -14,101 +12,17 @@ import { paymentType } from "../../utils/paymentType";
 import { DateUtils, extractFeeMedium, getAuthorizedUuid, getTaskType } from "@egovernments/digit-ui-module-dristi/src/Utils";
 import {
   formatRespondentAddressLine,
-  getPaymentModalDeliveryNoteDeadlineFormatted,
   getViewOrderClickHandler,
+  PaymentSummonFeeSelectionComponent,
   useCaseLockStatusForPaymentModal,
   useIsUserAdvocateOnCase,
 } from "./shared/paymentSummonModalShared";
-import ButtonSelector from "@egovernments/digit-ui-module-dristi/src/components/ButtonSelector";
 import { getPartyNameForInfos } from "../../utils";
 import { CaseWorkflowState } from "@egovernments/digit-ui-module-dristi/src/Utils/caseWorkflow";
 import { ORDER_TYPES } from "../../utils/constants";
 import { submitModalInfoPostPayment as submitModalInfo } from "./shared/paymentSummonPostReceiptDefaults";
 
 const modeOptions = [{ label: "E-Post (3-5 days)", value: "e-post" }];
-
-const PaymentForSummonComponent = ({
-  infos,
-  links,
-  feeOptions,
-  orderDate,
-  paymentLoader,
-  orderType,
-  isUserAdv,
-  isCaseLocked = false,
-  payOnlineButtonTitle = null,
-}) => {
-  const { t } = useTranslation();
-  const CustomErrorTooltip = window?.Digit?.ComponentRegistryService?.getComponent("CustomErrorTooltip");
-
-  const [selectedOption, setSelectedOption] = useState(modeOptions[0]);
-
-  return (
-    <div className="payment-for-summon">
-      <InfoCard
-        variant={"warning"}
-        label={"Complete in 2 days"}
-        additionalElements={[
-          <p>
-            {t(orderType === ORDER_TYPES.SUMMONS ? "SUMMON_DELIVERY_NOTE" : "NOTICE_DELIVERY_NOTE")}{" "}
-            <span style={{ fontWeight: "bold" }}>{getPaymentModalDeliveryNoteDeadlineFormatted()}</span> {t("ON_TIME_DELIVERY")}
-          </p>,
-        ]}
-        inline
-        textStyle={{}}
-        className={`custom-info-card warning`}
-      />
-      <CardLabel className="case-input-label">{t("OFFLINE_PAYMENT_INFO")}</CardLabel>
-      <ApplicationInfoComponent infos={infos} links={links} />
-      <LabelFieldPair className="case-label-field-pair">
-        <div className="join-case-tooltip-wrapper">
-          <CardLabel className="case-input-label">{t("Select preferred mode of post to pay")}</CardLabel>
-          <CustomErrorTooltip message={t("Select date")} showTooltip={true} icon />
-        </div>
-        <RadioButtons
-          additionalWrapperClass="mode-of-post-pay"
-          options={modeOptions}
-          selectedOption={selectedOption}
-          optionsKey={"label"}
-          onSelect={(value) => setSelectedOption(value)}
-          disabled={paymentLoader}
-        />
-      </LabelFieldPair>
-      {selectedOption?.value && (
-        <div className="summon-payment-action-table">
-          {feeOptions[selectedOption?.value]?.map((action, index) => (
-            <div className={`${index === 0 ? "header-row" : "action-row"}`}>
-              <div className="payment-label">{t(action?.label)}</div>
-              <div className="payment-amount">{action?.action !== "offline-process" && action?.amount ? `Rs. ${action?.amount}/-` : "-"}</div>
-              {isUserAdv && (
-                <div className="payment-action">
-                  {index === 0 ? (
-                    t(action?.action)
-                  ) : action?.isCompleted ? (
-                    <p style={{ color: "green" }}>{t("PAYMENT_COMPLETED")}</p>
-                  ) : action?.action !== "offline-process" ? (
-                    <ButtonSelector
-                      style={{ border: "1px solid" }}
-                      label={t(action.action)}
-                      onSubmit={action.onClick}
-                      isDisabled={paymentLoader || isCaseLocked}
-                      title={isCaseLocked ? t(payOnlineButtonTitle) : ""}
-                      textStyles={{ margin: "0px" }}
-                    />
-                  ) : (
-                    <p className="offline-process-text">
-                      {t("THIS_OFFLINE_TEXT")} <span className="learn-more-text">{t("LEARN_MORE")}</span>
-                    </p>
-                  )}
-                </div>
-              )}
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-};
 
 const PaymentForSummonModal = ({ path }) => {
   const history = useHistory();
@@ -449,10 +363,6 @@ const PaymentForSummonModal = ({ path }) => {
     ];
   }, [compositeItem, filteredTasks, hearingsData?.HearingList, orderDetails, orderType]);
 
-  const orderDate = useMemo(() => {
-    return hearingsData?.HearingList?.[0]?.startTime;
-  }, [hearingsData?.HearingList]);
-
   const links = useMemo(() => {
     return [{ text: "View order", link: "", onClick: getViewOrderClickHandler({ history, caseData, filingNumber }) }];
   }, [caseData?.criteria, filingNumber, history]);
@@ -472,21 +382,23 @@ const PaymentForSummonModal = ({ path }) => {
       payOnlineButtonTitle: payOnlineButtonTitle,
       className: "payment-modal",
       modalBody: (
-        <PaymentForSummonComponent
+        <PaymentSummonFeeSelectionComponent
           infos={infos}
           links={links}
           feeOptions={feeOptions}
-          orderDate={orderDate}
           paymentLoader={paymentLoader}
-          isCaseAdmitted={isCaseAdmitted}
           orderType={orderType}
           isUserAdv={isUserAdv}
           isCaseLocked={isCaseLocked}
           payOnlineButtonTitle={payOnlineButtonTitle}
+          modeOptions={modeOptions}
+          modeSelectionLabelKey="Select preferred mode of post to pay"
+          showOfflinePaymentInfoCardLabel={true}
+          offlineHelpVariant="epost-simple"
         />
       ),
     };
-  }, [feeOptions, history, infos, isCaseAdmitted, links, orderDate, orderType, paymentLoader, isUserAdv]);
+  }, [feeOptions, history, infos, links, orderType, paymentLoader, isUserAdv, isCaseLocked, payOnlineButtonTitle]);
 
   if (isOrdersLoading || !orderData || isSummonsBreakUpLoading || isCourtBillLoading || isTaskLoading || isHearingLoading) {
     return <Loader />;
